requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
enum34==1.1.6
flake8==3.5.0
flake8-polyfill==1.0.2
h5py==2.7.0
html5lib==0.9999999
idna==2.6
Keras==2.0.6
Markdown==2.6.11
mccabe==0.6.1
mypy==0.560
nose==1.3.7
numpy==1.13.1
pep8-naming==0.5.0
Pillow==5.0.0
protobuf==3.5.1
psutil==5.4.3
pycodestyle==2.3.1
pyflakes==1.6.0
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorflow==1.2.1
tensorflow-tensorboard==0.4.0rc3
Theano==1.0.1
typed-ast==1.1.0
urllib3==1.22
Werkzeug==0.14.1
